feat(run)!: pure rig_run::prepare_request; RequestPatch moves to rig_run::policy
Request preparation is protocol, not driver: given a RunSpec, the model's
capability snapshot, the history, the tools available this turn, the run's
committed output tool and a per-turn RequestPatch, exactly one request
follows. rig_run::prepare::prepare_request is that function — sync, IO-free
(the no-async guard still passes) — returning a PreparedRequest of owned
data (chat history with the augmented preamble as the leading system
message, static + extra documents, advertised tools incl. the synthetic
output tool, temperature/max_tokens/additional_params/tool_choice, the
native output schema only in Native mode, the resolved OutputMode,
executable/allowed tool names, output_tool_name) plus
PreparedRequest::apply(builder) to bind it to any CompletionRequestBuilder<M>
in the canonical order. PrepareError carries the same local, pre-IO error
messages as before and converts into CompletionError::RequestError.

The output-mode resolution, output-tool naming/synthesis, tool-choice
validation and their unit tests move verbatim from rig-agent's completion.rs
into prepare.rs. RequestPatch (plain hook data: is_empty/merge now public)
moves from rig_agent::agent::hook to rig_run::policy; the hook that
produces it stays in rig-agent, which re-exports RequestPatch at the old
paths.

rig-agent's build_prepared_completion_request is now the driver's share
only: retrieve this turn's tools (the one .await), prepare_request, pin the
snapshot to the executable set, bind the selected model's builder, set the
telemetry flag. It also hands the run the tools the request actually
carries (AgentRun::advertise_tools) — previously the advertised list was
taken after the definitions had been moved out of the snapshot, so it was
always empty.

Identity gate: a new golden test (completion.rs request_identity_tests)
pins the full serialized CompletionRequests of a scripted tool turn with
preamble, static context, output schema in Tool mode and a CompletionCall
patch (preamble/temperature/max_tokens/tool_choice/active_tools/
additional_params/extra_context); the same test passes unchanged against
the pre-move driver on main. Whole-binary cassette suites (anthropic 420,
openai 420, gemini 348, groq 82) replay green against their recorded
request bodies.
